analytics_service/requirements.txt
kafka-python==2.0.3
pandas==1.4.3
numpy==1.22.0
SQLAlchemy
psycopg2-binary
